fix: support file urls in patchedFs
