+++ packages/yarnpkg-fslib/tests/patchedFs.test.ts
@@ -1,14 +1,16 @@
-import fs                       from 'fs';
-import {promisify}              from 'util';
+import fs                from 'fs';
+import {pathToFileURL}   from 'url';
+import {promisify}       from 'util';
 
-import {NodeFS}                 from '../sources/NodeFS';
-import {PosixFS}                from '../sources/PosixFS';
-import {Filename, npath, ppath} from '../sources/path';
-import {extendFs}               from '../sources';
+import {NodeFS}          from '../sources/NodeFS';
+import {PosixFS}         from '../sources/PosixFS';
+import {extendFs}        from '../sources/patchFs';
+import {Filename, npath} from '../sources/path';
+import {xfs}             from '../sources/xfs';
 
 describe(`patchedFs`, () => {
   it(`in case of no error, give null: fs.stat`, done => {
-    const file = ppath.join(npath.toPortablePath(__dirname), `patchedFs.test.ts` as Filename);
+    const file = npath.join(__dirname, `patchedFs.test.ts` as Filename);
 
     const patchedFs = extendFs(fs, new PosixFS(new NodeFS()));
 
@@ -19,7 +21,7 @@ describe(`patchedFs`, () => {
   });
 
   it(`in case of no error, give null: fs.read`, done => {
-    const file = ppath.join(npath.toPortablePath(__dirname), `patchedFs.test.ts` as Filename);
+    const file = npath.join(__dirname, `patchedFs.test.ts` as Filename);
 
     const patchedFs = extendFs(fs, new PosixFS(new NodeFS()));
 
@@ -45,7 +47,7 @@ describe(`patchedFs`, () => {
     const patchedFs = extendFs(fs, new PosixFS(new NodeFS()));
     const patchedFsReadAsync = promisify(patchedFs.read);
 
-    const file = ppath.join(npath.toPortablePath(__dirname), `patchedFs.test.ts` as Filename);
+    const file = npath.join(__dirname, `patchedFs.test.ts` as Filename);
 
     const fd = fs.openSync(file, `r`);
 
@@ -55,5 +57,31 @@ describe(`patchedFs`, () => {
 
     expect(typeof result.bytesRead).toBe(`number`);
     expect(Buffer.isBuffer(result.buffer)).toBeTruthy();
+  });
+
+  it(`should support URL instances`, () => {
+    const patchedFs = extendFs(fs, new PosixFS(new NodeFS()));
+
+    const tmpdir = npath.fromPortablePath(xfs.mktempSync());
+    const tmpdirUrl = pathToFileURL(tmpdir);
+
+    const file = `${tmpdir}/file.txt`;
+    const fileUrl = pathToFileURL(file);
+
+    patchedFs.writeFileSync(fileUrl, `Hello World`);
+
+    expect(patchedFs.readdirSync(tmpdirUrl)).toStrictEqual(patchedFs.readdirSync(tmpdir));
+
+    expect(patchedFs.readFileSync(fileUrl, {encoding: `utf8`})).toStrictEqual(patchedFs.readFileSync(file, {encoding: `utf8`}));
+    expect(patchedFs.statSync(fileUrl)).toStrictEqual(patchedFs.statSync(file));
+
+    const copyUrl = pathToFileURL(`${tmpdir}/copy.txt`);
+    const renamedUrl = pathToFileURL(`${tmpdir}/renamed.txt`);
+
+    patchedFs.copyFileSync(fileUrl, copyUrl);
+    patchedFs.renameSync(copyUrl, renamedUrl);
+    patchedFs.unlinkSync(renamedUrl);
+
+    expect(patchedFs.existsSync(renamedUrl)).toStrictEqual(false);
   });
 });
